+++ demo/pages/_app.tsx
@@ -24,7 +24,7 @@ import {
   ReservoirKitTheme,
   CartProvider,
 } from '@reservoir0x/reservoir-kit-ui'
-import { LogLevel } from '@reservoir0x/reservoir-sdk'
+import { LogLevel, customChains } from '@reservoir0x/reservoir-sdk'
 import configuredChains from '../utils/chains'
 import { useRouter } from 'next/router'
 import '../fonts.css'
@@ -54,7 +54,8 @@ const { chains, publicClient } = configureChains(
     allChains.avalanche,
     allChains.linea,
     allChains.scroll,
-    allChains.arbitrumNova
+    allChains.arbitrumNova,
+    customChains.frameTestnet
   ],
   [alchemyProvider({ apiKey: ALCHEMY_KEY }), publicProvider()]
 )