+++ packages/ui/src/constants/chainIcons.tsx
@@ -43,6 +43,9 @@ import ZksyncIconColor from '../img/chains/ZksyncIconColor'
 import ZoraIconDark from '../img/chains/ZoraIconDark'
 import ZoraIconLight from '../img/chains/ZoraIconLight'
 import ZoraIconColor from '../img/chains/ZoraIconColor'
+import FrameIconDark from '../img/chains/FrameIconDark'
+import FrameIconLight from '../img/chains/FrameIconLight'
+import FrameIconColor from '../img/chains/FrameIconColor'
 
 const chainIcons = {
   [reservoirChains.ancient8Testnet.id]: {
@@ -139,6 +142,11 @@ const chainIcons = {
     light: <ZoraIconDark />,
     dark: <ZoraIconLight />,
     color: <ZoraIconColor />,
+  },
+  [reservoirChains.frameTestnet.id]: {
+    light: <FrameIconDark />,
+    dark: <FrameIconLight />,
+    color: <FrameIconColor />,
   },
 }