+++ packages/ui/src/modal/cancelBid/CancelBidModalRenderer.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useEffect, useState, useCallback, ReactNode } from 'react'
 import { useCoinConversion, useReservoirClient, useBids } from '../../hooks'
 import { useWalletClient } from 'wagmi'
-import { Execute, ReservoirWallet } from '@reservoir0x/reservoir-sdk'
+import { Execute, ReservoirWallet, axios } from '@reservoir0x/reservoir-sdk'
 import { getNetwork, switchNetwork } from 'wagmi/actions'
 import { customChains } from '@reservoir0x/reservoir-sdk'
 import * as allChains from 'viem/chains'
@@ -204,6 +204,9 @@ export const CancelBidModalRenderer: FC<Props> = ({
       setTransactionError(null)
       setStepData(null)
       setSteps(null)
+      axios.defaults.headers.common['x-rkui-context'] = ''
+    } else {
+      axios.defaults.headers.common['x-rkui-context'] = 'cancelBidModalRenderer'
     }
   }, [open])